dev-requirements.txt
flake8==6.0.0
black==23.3.0
mypy==1.2.0
isort==5.12.0
types-requests==2.27.1
types-setuptools==57.4.3
types-psutil==5.8.0
grpcio-tools==1.43.0
types-dataclasses==0.6.6
backoff-stubs~=1.10
pytest~=7.0.1
pytest-asyncio~=0.21.0
types-beautifulsoup4==4.11.1
types-python-dateutil~=2.8.2
requests-mock~=1.10.0
